chatbot-service/pyproject.toml
[project]
name = "chatbot-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "py-eureka-client>=0.11.13",
    "python-dotenv>=1.1.0",
    "uvicorn[standard]>=0.34.3",
]
